Refactor Firebase and authentication modules to streamline logging integration and improve code organization
